New Old Attributes?
New Old Attributes(NOAttr?) is a way to bring back old legacy attributes with the use of CSS. It is extremely helpful for people who have worked with the attributes: bgcolor, color, border, width, and height. It is used by adding custom attributes via CSS and the "attr()" function, which grabs an attribute's value for the element and uses it with the CSS element to recreate the old attributes.